clean up types
