+++ backend/src/routes/identify.ts
@@ -1,7 +1,6 @@
 import express, { Request, Response } from 'express';
 import { MovieIDManager } from '../controllers/MovieIDManager';
-import { Input } from '../types/types';
-import { ForumResponse, RequestMoreInformation } from '../blackboard/Forum';
+import { Input, ForumResponse, RequestMoreInformation } from '../types/types';
 
 const router = express.Router();
 const movieIDManager = new MovieIDManager();
