+++ backend/src/types/types.ts
@@ -22,3 +22,29 @@ export type AudioInput = {
 	type: "audio";
 	data: string;
 }
+
+export type StoredResponse = {
+	userId: string;
+	forumResponse : ForumResponse;
+}
+
+export type ForumResponse = {
+	responseId: string;
+	overallConfidence: number;
+	movieName: string;
+	timestamp: number;
+	inputsUsed: string[];
+}
+
+export type RequestMoreInformation = {
+	inputsUsed: string[];
+	details: string;
+}
+
+export interface ExpertResponse {
+  expertName: string;
+  movies: string[];
+  confidence: number;
+  timestamp?: number;
+  details?: String;
+}
